+++ src/types.ts
@@ -233,3 +233,19 @@ export type validColors =
     | "italic"
     | "bright-blue"
     | "bright-yellow";
+
+/**
+ * Possible intensities of all cleaners.
+ *
+ * @export
+ * @typedef {CleanerIntensity}
+ */
+export type CleanerIntensity = "normal" | "hard" | "hard-only" | "maxim";
+
+/**
+ * Possible intensities of the project cleaner.
+ *
+ * @export
+ * @typedef {ProjectCleanerIntensity}
+ */
+export type ProjectCleanerIntensity = "normal" | "hard" | "maxim";